total text, axis line
Total text: If calc_total is TRUE, 'total_axis_text' can display a
custom text string on the axis; before, it was always "Total"

Axis line: The axis line now defaults to be black (instead of
invisible). It can be set behind the recangle (default), in front of the
rectangle (if that's the way you want to live your life), or it can be
omitted as before.