Fix bug in temp budget repair levy
The function `ifelse` returns arguments of the same length as `cond`, irrespective of the lengths of `yes` and `no`. Consequently, if `fy.year` had length 1, the temporary budget repair levy would be calculated on based on `income[1]`.

This bug was picked up during a test of `inverse_income_tax` earlier this morning. This commit should close #19.